ADDED IMU TO LS FIRMWARE AND CHANGED GS FIRMWARE
Added Adafruit BNO055 support to the launchsonde firmware in v0.1.3. This code is confirmed to work and transmit IMU and altitude data over the LoRa radio and locally log on the SD Card.

The ground station firmware has also been pushed to a new alpha version, v0.1. This version supports the verbose output of the launchsonde firmware over the LoRa radio.
Logging needs to be done on the host computer in .CSV format.